requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
sqlalchemy==2.0.18
pydantic==1.11.1
requests==2.31.0
python-dotenv==0.21.0
psycopg2==2.9.5  
alembic==1.9.3 
clerk
python-jose
asyncio

